feat: Fix Zipper module
